+++ src/README.md
@@ -8,7 +8,8 @@ This folder contains several LightningDataModule and LightningModule classes.
 
 ## LightningModule (model architecture)
 
-- model_vit.py - Vision Transformer neural network model architecture
+- model_clay.py - Clay Foundation Model architecture with spatiotemporal encoders
+- model_vit.py - Vanilla Vision Transformer neural network model architecture
 
 ## References
