+++ src/model_clay.py
@@ -1,3 +1,4 @@
+import lightning as L
 import torch
 import torch.nn.functional as F
 from einops import rearrange, reduce, repeat
@@ -692,3 +693,51 @@ def clay_large(**kwargs):
     }
     model = CLAY(**args.update(kwargs))
     return model
+
+
+class CLAYModule(L.LightningModule):
+    def __init__(self, lr=1e-4, wd=0.05, b1=0.9, b2=0.95):
+        super().__init__()
+        self.save_hyperparameters(logger=True)
+        self.model = clay_small(lr=lr, wd=wd, b1=b1, b2=b2)
+
+    def forward(self, cube: dict[str, torch.Tensor]):
+        return self.model(cube)
+
+    def configure_optimizers(self):
+        optimizer = torch.optim.AdamW(
+            self.parameters(),
+            lr=self.hparams.lr,
+            weight_decay=self.hparams.wd,
+            betas=(self.hparams.b1, self.hparams.b2),
+        )
+        scheduler = torch.optim.lr_scheduler.CosineAnnealingWarmRestarts(
+            optimizer, T_0=200, T_mult=2, eta_min=self.hparams.lr * 10, last_epoch=-1
+        )
+
+        return {
+            "optimizer": optimizer,
+            "lr_scheduler": {
+                "scheduler": scheduler,
+                "interval": "step",
+            },
+        }
+
+    def shared_step(self, batch: dict[str, torch.Tensor], batch_idx: int, phase: str):
+        cube = batch
+        loss = self(cube)
+        self.log(
+            name=f"{phase}/loss",
+            value=loss,
+            on_step=True,
+            on_epoch=True,
+            prog_bar=True,
+            logger=True,
+        )
+        return loss
+
+    def training_step(self, batch: dict[str, torch.Tensor], batch_idx: int):
+        return self.shared_step(batch, batch_idx, phase="train")
+
+    def validation_step(self, batch: dict[str, torch.Tensor], batch_idx: int):
+        return self.shared_step(batch, batch_idx, phase="val")
